+++ tomcat/src/main/java/org/apache/coyote/http11/application/Handler.java
@@ -6,7 +6,7 @@ import org.apache.coyote.http11.request.Api;
 import org.apache.coyote.http11.request.HttpRequest;
 import org.apache.coyote.http11.response.HttpResponse;
 
-public interface Controller {
+public interface Handler {
 
     Function<HttpRequest, HttpResponse> getHandlerMethod(Api requestApi);
 }
